+++ Cotacoes.Model/conversao/ConversaoService.cs
@@ -2,7 +2,7 @@ namespace Cotacoes.Model
 {
     public static class ConversaoService
     {
-        public static Conversao ConverterParaReais(double Montante, string SiglaMoeda)
+        public static Conversao ConverterParaReais(decimal Montante, string SiglaMoeda)
         {
             if (!string.IsNullOrEmpty(SiglaMoeda))
             {
@@ -13,7 +13,7 @@ namespace Cotacoes.Model
                 throw new System.Exception("Erro ao efetuar a conversão, verifique a sigla");
             }
         }
-        public static Conversao ConverterParaDolar(double Montante, string SiglaMoeda)
+        public static Conversao ConverterParaDolar(decimal Montante, string SiglaMoeda)
         {
             if (!string.IsNullOrEmpty(SiglaMoeda))
             {
